01CloudFunctions/a1_getcookies/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage
pandas >= 1.1.0
python-dateutil==2.8.1
numpy==1.21.1
DateTime==4.3
requests